initial win32 port

+++ apps/texrecon/texrecon.cpp
@@ -150,7 +150,7 @@ int main(int argc, char **argv) {
         } else {
             ProgressCounter texture_patch_counter("Calculating validity masks for texture patches", texture_patches.size());
             #pragma omp parallel for schedule(dynamic)
-            for (std::size_t i = 0; i < texture_patches.size(); ++i) {
+            for (std::int64_t i = 0; i < texture_patches.size(); ++i) {
                 texture_patch_counter.progress<SIMPLE>();
                 TexturePatch::Ptr texture_patch = texture_patches[i];
                 std::vector<math::Vec3f> patch_adjust_values(texture_patch->get_faces().size() * 3, math::Vec3f(0.0f));
