+++ libs/tex/generate_texture_views.cpp
@@ -114,7 +114,7 @@ from_images_and_camera_files(std::string const & path, std::vector<TextureView> 
 
     ProgressCounter view_counter("\tLoading", files.size() / 2);
     #pragma omp parallel for
-    for (std::size_t i = 0; i < files.size(); i += 2) {
+    for (std::int64_t i = 0; i < files.size(); i += 2) {
         view_counter.progress<SIMPLE>();
         std::string cam_file = files[i];
         std::string img_file = files[i + 1];
@@ -189,7 +189,7 @@ from_nvm_scene(std::string const & nvm_file, std::vector<TextureView> * texture_
 
     ProgressCounter view_counter("\tLoading", cameras.size());
     #pragma omp parallel for
-    for (std::size_t i = 0; i < cameras.size(); ++i) {
+    for (std::int64_t i = 0; i < cameras.size(); ++i) {
         view_counter.progress<SIMPLE>();
         mve::CameraInfo& mve_cam = cameras[i];
         mve::NVMCameraInfo const& nvm_cam = nvm_cams[i];
